+++ engine-vuex/CHANGELOG.md
@@ -1,10 +1,17 @@
+# @wwtelescope/engine-vuex 0.13.0 (2021-11-17)
+
+- Expose the current roll angle of the camera view in `WWTAwareComponent` (#155,
+  @Carifio24).
+- Fix a copy/paste-o in `spreadSheetLayerByKey` error message (#155,
+  @Carifio24).
+
 # @wwtelescope/engine-vuex 0.12.0 (2021-10-25)
 
 - Add new interfaces that help unify code that deals with both HiPS
   progressive catalog layers and regular "spreadsheet" layers (#146,
   @Carifio24). These include the `CatalogLayerInfo` and `SpreadSheetLayerInfo`
   classes and `WWTEngineVuexModule` getters `spreadSheetLayerById`,
-  `spreadsheetStateById`, `spreadSheetLayer`, and `spreadsheetState`; 
+  `spreadsheetStateById`, `spreadSheetLayer`, and `spreadsheetState`;
 
 
 # @wwtelescope/engine-vuex 0.11.0 (2021-09-20)