+++ research-app-messages/CHANGELOG.md
@@ -1,3 +1,15 @@
+# @wwtelescope/research-app-messages 0.12.0 (2021-11-17)
+
+New APIs to help support startup scripting of the research app (#155,
+@Carifio24):
+
+- `layers.isLoadHipsCatalogCompletedMessage`
+- `layers.MultiModifyTableLayerMessage` and type guard function
+- `layers.MultiModifyFitsLayerMessage` and type guard function
+- `layers.MultiModifyAnnotationMessage` and type guard function
+- `selections.AddSourceMessage` and type guard function
+
+
 # @wwtelescope/research-app-messages 0.11.2 (2021-10-27)
 
 - Not quite "fix the fix", but ... improve it (#154, @pkgw). We only replaced